requirements.txt
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.27
langchainhub==0.1.21
python-dotenv==1.1.1
faker>=22.0.0
duckdb>=0.9.0
langfuse>=2.0.0
python-a2a>=0.5.9
